feat(login): fold logout into login as a contextual remove action
system2 login now handles both flows in one command. When the user
selects a provider that's already logged in, a 3-way menu offers
re-login, remove, or cancel. The remove action deletes the credential
file and removes the provider from [llm.oauth] in one step (the
previous logout flow asked for two confirms).

Removes:
- src/cli/commands/logout.ts (file deleted, helper moved to login.ts)
- src/cli/commands/logout.test.ts (tests merged into login.test.ts)
- system2 logout CLI command registration

Helps to avoid the second hardcoded OAUTH_PROVIDERS list in logout.ts
that would otherwise have needed expanding to all 5 providers, and
collapses the conceptual surface from "login + logout" to a single
"manage OAuth providers" command.

Migration: users who had `system2 logout <provider>` in scripts must
either run `system2 login` interactively, or edit ~/.system2/config.toml
and rm ~/.system2/oauth/<provider>.json directly.

Also: biome auto-organized imports in src/cli/utils/config.ts and
src/cli/utils/config.test.ts.
